moved search as index
