+++ pages/cases/index.jsx
@@ -1,14 +1,12 @@
 import { NextSeo } from 'next-seo';
 
-import BackButton from 'components/Layout/BackButton/BackButton';
 import Search from 'components/Search/Search';
 
-const SearchResidentPage = ({ query }) => {
+const SearchCasesPage = ({ query }) => {
   return (
     <div>
-      <NextSeo title="Search" noindex />
-      <BackButton />
-      <Search query={query} type="people" />
+      <NextSeo title="Find unlinked case notes" noindex />
+      <Search query={query} type="cases" />
     </div>
   );
 };
@@ -22,4 +20,4 @@ export const getServerSideProps = async (ctx) => {
   };
 };
 
-export default SearchResidentPage;
+export default SearchCasesPage;
